Connect socket within GameRouter
