+++ frontend/src/GameRouter.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useContext, useEffect, useMemo, useState } from "react";
 import { Redirect, Route, Router, Switch, useParams } from "react-router-dom";
 import { createMemoryHistory } from "history";
 import {
@@ -13,6 +13,9 @@ import {
 import { SocketProvider } from "./contexts/socket";
 import io from "socket.io-client";
 import { useSetupSocketHandlers } from "./hooks/socket";
+import createPersistedState from "use-persisted-state";
+import { BrowserHistoryContext } from "./App";
+import { useGet } from "./hooks/axios";
 
 const socket = io({
   autoConnect: false,
@@ -32,14 +35,60 @@ const memoryHistory = createMemoryHistory();
 const INITIAL_ROUND_LIMIT = 69;
 const INITIAL_MAX_PLAYERS = 40;
 
+const usePlayerIdState = createPersistedState("playerId");
+const useTokenState = createPersistedState("token");
+
 const GameRouter = () => {
+  const [playerId] = usePlayerIdState("");
+  const [token] = useTokenState("");
+
   const { gameCode } = useParams<PathParams>();
   const [settings, setSettings] = useState<Settings>({
     roundLimit: INITIAL_ROUND_LIMIT,
     maxPlayers: INITIAL_MAX_PLAYERS,
   });
+  const browserHistory = useContext(BrowserHistoryContext);
+
+  const [, validateGameCode] = useGet(
+    "/api/game/validate",
+    /*
+    As `validateGameCode` runs inside a `useEffect()` we need to create the
+    config object with `useMemo()` to prevent unnecessary re-renders.
+     */
+    useMemo(
+      () => ({
+        params: { gameCode },
+      }),
+      [gameCode]
+    )
+  );
+
+  useEffect(() => {
+    const controller = new AbortController();
+    (async () => {
+      const response = await validateGameCode(controller);
+      if (response === null) {
+        browserHistory.push("/");
+      }
+    })();
+    return () => controller.abort();
+  }, [browserHistory, validateGameCode]);
 
   useSetupSocketHandlers(socket, memoryHistory, settings, setSettings);
+
+  useEffect(() => {
+    if (playerId && token) {
+      socket.auth = { gameCode, playerId, token };
+      /*
+      If the connection fails here, `playerId` and `token` are cleared by the
+      `connect_error` event handler initialised in `useSetupSocketHandlers()`.
+       */
+      socket.connect();
+    }
+    return () => {
+      socket.close();
+    };
+  }, [gameCode, playerId, token]);
 
   return (
     <SocketProvider socket={socket}>
